Move code up to app to allow reseting to be easier
